LSG-CLASS: Fix RPM uninstallation on CBL-Mariner
CBL-Mariner doesn't provide a kernel tool used by the generic RPM spec
file generator to properly uninstall the kernel.  This patch fixes the
uninstallation step to update the /boot/mariner.cfg symlink [1].

Because this patch is specific to this Linux distribution, it doesn't
make sense to send it upstream.  A better fix would require CBL-Mariner
to provide the update-bootloader tool and adapt installkernel, but this
is a kernel class. :)

Anyway, this is a good opportunity to apply a first kernel patch.

Link: https://github.com/microsoft/CBL-Mariner/blob/1.0/SPECS/kernel/kernel.spec#L543-L552 [1]
Signed-off-by: Mickaël Salaün <misalaun@microsoft.com>